Turn the model picker's "More models" into an expandable list of hover submenus
Clicking "More models" expands the maker list inline, in place. Each
provider is a submenu that opens on hover, showing its model list —
the dropdown's standard, native submenu behavior, no custom navigation
state or transition needed.
